+++ src/frontend/pages/municipalities/index.vue
@@ -1,20 +1,5 @@
-<template>
-  <div>
-    <div class="prose mb-8 mt-10 max-w-full text-center">
-      <h1 class="mb-0 whitespace-pre-line">
-        {{ $t("municipalities.heading") }}
-      </h1>
-      <p class="mt-0 text-xs">
-        {{ $t("municipalities.last_updated_at") + lastUpdatedAtStr }}
-      </p>
-    </div>
-    <section>
-      <the-ranking :municipalities="municipalities"></the-ranking>
-    </section>
-  </div>
-</template>
-
 <script setup>
+import { ref } from 'vue'
 import lodash from "lodash";
 const { sortBy, last, get } = lodash;
 const { $directus, $readItems, $t, $locale } = useNuxtApp();
@@ -24,11 +9,12 @@ const title = ref($t("municipalities.nav_label"));
 useHead({
   title,
 });
-//
+
+// Fetch all published municipalities from directus
 const { data: municipalities } = await useAsyncData("municipalities", () => {
   return $directus.request(
     $readItems("municipalities", {
-      fields: ["slug", "name", "score_total", "place", "state", "date_updated"],
+      fields: ["slug", "name", "score_total", "place", "state", "date_updated", "municipality_type"],
       sort: "-score_total",
       limit: -1,
       filter: {
@@ -40,6 +26,19 @@ const { data: municipalities } = await useAsyncData("municipalities", () => {
   );
 });
 
+// todo fix "place" for these views
+const cities = getSublist((municipality) => municipality.municipality_type === 'big_city');
+const towns = getSublist((municipality) => municipality.municipality_type === 'small_city');
+
+function getSublist(condition) {
+  return (municipalities.value?.filter(condition) || [])
+    .map((item, index) => ({
+      ...item,
+      place: index + 1,
+    }));
+}
+
+
 const lastUpdatedAtStr = ref("");
 onMounted(() => {
   const lastUpdatedAt = new Date(get(last(sortBy(municipalities.value, ["date_updated"])), "date_updated"));
@@ -48,4 +47,63 @@ onMounted(() => {
     ", " +
     lastUpdatedAt.toLocaleTimeString($locale);
 });
+
+// Toggle between "grossstaedte" and "gemeinden"
+const selected = ref('city')
+
+import citySelected from '~/assets/images/city-light.svg'
+import cityNotSelected from '~/assets/images/city-dark.svg'
+import townSelected from '~/assets/images/town-light.svg'
+import townNotSelected from '~/assets/images/town-dark.svg'
+// TODO - translations
 </script>
+
+<template>
+
+  <div>
+    <div class="prose mb-8 mt-10 max-w-full text-center">
+      <h1 class="mb-0 whitespace-pre-line">
+        {{ $t("municipalities.heading") }}
+      </h1>
+      <p class="mt-0 text-xs">
+        {{ $t("municipalities.last_updated_at") + lastUpdatedAtStr }}
+      </p>
+    </div>
+  </div>
+
+  <!-- Button row -->
+  <div class="flex flex-row justify-center gap-2 w-full px-4 sm:px-0 sm:w-auto sm:gap-4">
+    <!-- Major cities -->
+    <div class="flex flex-col items-center w-1/2 sm:w-[22%] max-w-[240px]">
+      <img :src="selected === 'city' ? citySelected : cityNotSelected" alt="Großstädte"
+        class="cursor-pointer w-full h-auto" @click="selected = 'city'" />
+      <p v-if="selected === 'city'" class="text-xs text-center mt-1">
+        mehr als 100.000 Menschen
+      </p>
+    </div>
+
+    <!-- Small cities -->
+    <div class="flex flex-col items-center w-1/2 sm:w-[22%] max-w-[240px]">
+      <img :src="selected === 'town' ? townSelected : townNotSelected" alt="Städte und Gemeinden"
+        class="cursor-pointer w-full h-auto" @click="selected = 'town'" />
+      <p v-if="selected === 'town'" class="text-xs text-center mt-1">
+        weniger als 100.000 Menschen
+      </p>
+    </div>
+  </div>
+
+
+
+  <!-- Conditional Content -->
+  <div v-if="selected === 'city'">
+    <section>
+      <the-ranking :municipalities="cities"></the-ranking>
+    </section>
+  </div>
+  <div v-else>
+    <section>
+      <the-ranking :municipalities="towns"></the-ranking>
+    </section>
+  </div>
+
+</template>
